RDMA/hfi1: Add port to file and send context structs
The hfi1 file data structure is directly associated with a
port.  Send contexts are assigned on a per port basis.  Add
a port reference to these structures for later use.

This causes a ripple effect in calling parameters, including:
o init_per_vl_scs(): Move to be initialized on a per-port basis.
  This is fine since, presently, there is only one port.
o allocate_ctxt(): Remove a redundant call parameter.

Signed-off-by: Dean Luick <dean.luick@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>