RDMA/core: Add writev to uverbs file descriptor
Add a writev pass-through between the uverbs file descriptor and
infiniband devices.  Interested devices may subscribe to this
functionality.

The goal is to keep all the semantics of the user interface the same so
it's an easy migration to the uverbs cdev from the private cdev. The idea
is that all the command and control is still ioctl, but the "data path" is
still using the writev() to pass in the iovecs.

Signed-off-by: Dean Luick <dean.luick@cornelisnetworks.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>

---
Changes since v1:
Updated commit message to indiate why we are keeping the
writev semantic.