RDMA/qib: Remove outdated driver
The qib Infiniband device has long been decommissioned from distros and
unsupported by the company. We have kept it going in a compile only mode
for a while now and it's time to remove the driver. It has a number of
issues that are never going to be addressed and is a hinderance to
improving hfi1 and rdmavt. 

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>